Move most logic to separate file
